+++ component/pl_esp_network_interface.cpp
@@ -13,8 +13,10 @@ namespace PL {
 //==============================================================================
 
 EspNetworkInterface::~EspNetworkInterface() {
-  if (netif)
-    esp_event_handler_instance_unregister(IP_EVENT, ESP_EVENT_ANY_ID, eventHandlerInstance);
+  if (eventHandlerInstance) {
+    ESP_LOGE(TAG, "UnregisterEventHandler was not called by the derived class destructor");
+    abort();
+  }
 }
 
 //==============================================================================
@@ -168,6 +170,16 @@ esp_err_t EspNetworkInterface::InitializeNetif(esp_netif_t* netif) {
   LockGuard lg(*this);
   ESP_RETURN_ON_ERROR(esp_event_handler_instance_register(IP_EVENT, ESP_EVENT_ANY_ID, EventHandler, this, &eventHandlerInstance), TAG, "event handler instance register failed");
   this->netif = netif;
+  return ESP_OK;
+}
+
+//==============================================================================
+
+esp_err_t EspNetworkInterface::UnregisterEventHandler() {
+  if (!eventHandlerInstance)
+    return ESP_OK;
+  ESP_RETURN_ON_ERROR(esp_event_handler_instance_unregister(IP_EVENT, ESP_EVENT_ANY_ID, eventHandlerInstance), TAG, "event handler instance unregister failed");
+  eventHandlerInstance = NULL;
   return ESP_OK;
 }